+++ packages/svelte/src/compiler/compile/Component.js
@@ -770,14 +770,12 @@ export default class Component {
 			if (name[0] === '$') {
 				return this.error(/** @type {any} */ (node), compiler_errors.illegal_declaration);
 			}
-			const writable =
-				node.type === 'VariableDeclaration' && (node.kind === 'var' || node.kind === 'let');
-			const imported = node.type.startsWith('Import');
+			const { type } = node;
 			this.add_var(node, {
 				name,
 				initialised: instance_scope.initialised_declarations.has(name),
-				writable,
-				imported
+				imported: type.startsWith('Import'),
+				writable: type === 'VariableDeclaration' && (node.kind === 'var' || node.kind === 'let')
 			});
 			this.node_for_declaration.set(name, node);
 		});
@@ -1710,6 +1708,7 @@ function process_component_options(component, nodes) {
 					case 'customElement': {
 						component_options.customElement =
 							component_options.customElement || /** @type {any} */ ({});
+
 						const { value } = attribute;
 						if (value[0].type === 'MustacheTag' && value[0].expression?.value === null) {
 							component_options.customElement.tag = null;
@@ -1720,12 +1719,14 @@ function process_component_options(component, nodes) {
 						} else if (value[0].expression.type !== 'ObjectExpression') {
 							return component.error(attribute, compiler_errors.invalid_customElement_attribute);
 						}
+
 						const tag = value[0].expression.properties.find((prop) => prop.key.name === 'tag');
 						if (tag) {
 							parse_tag(tag, tag.value?.value);
 						} else {
 							return component.error(attribute, compiler_errors.invalid_customElement_attribute);
 						}
+
 						const props = value[0].expression.properties.find((prop) => prop.key.name === 'props');
 						if (props) {
 							const error = () =>
@@ -1770,6 +1771,7 @@ function process_component_options(component, nodes) {
 								}
 							}
 						}
+
 						const shadow = value[0].expression.properties.find(
 							(prop) => prop.key.name === 'shadow'
 						);
@@ -1780,6 +1782,14 @@ function process_component_options(component, nodes) {
 							}
 							component_options.customElement.shadow = shadowdom;
 						}
+
+						const extend = value[0].expression.properties.find(
+							(prop) => prop.key.name === 'extend'
+						);
+						if (extend?.value) {
+							component_options.customElement.extend = extend.value;
+						}
+
 						break;
 					}
 					case 'namespace': {
@@ -1853,7 +1863,8 @@ function get_sourcemap_source_filename(compile_options) {
 		: get_basename(compile_options.filename);
 }
 
-/** @typedef {Object} ComponentOptions
+/**
+ * @typedef {Object} ComponentOptions
  * @property {string} [namespace]
  * @property {boolean} [immutable]
  * @property {boolean} [accessors]
@@ -1862,4 +1873,5 @@ function get_sourcemap_source_filename(compile_options) {
  * @property {string|null} customElement.tag
  * @property {'open'|'none'} [customElement.shadow]
  * @property {Record<string,{attribute?:string;reflect?:boolean;type?:'String'|'Boolean'|'Number'|'Array'|'Object';}>} [customElement.props]
+ * @property {(ceClass: new () => HTMLElement) => new () => HTMLElement} [customElement.extend]
  */